requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
async-timeout==4.0.3
asyncssh==2.15.0
attrs==23.2.0
boto3==1.34.144
botocore==1.34.144
cffi==1.16.0
cryptography==42.0.8
discord==2.3.2
discord.py==2.4.0
exchange_calendars==4.5.5
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
korean-lunar-calendar==0.3.1
multidict==6.0.5
numpy==2.0.0
pandas==2.2.2
pandas_market_calendars==4.4.1
pika==1.3.2
pycparser==2.22
pyluach==2.2.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
yarl==1.9.4
